+++ magpylib/_src/display/traces_generic.py
@@ -43,6 +43,7 @@ from magpylib._src.display.traces_utility import group_traces
 from magpylib._src.display.traces_utility import merge_mesh3d
 from magpylib._src.display.traces_utility import merge_traces
 from magpylib._src.display.traces_utility import place_and_orient_model3d
+from magpylib._src.display.traces_utility import slice_mesh_from_colorscale
 from magpylib._src.display.traces_utility import triangles_area
 from magpylib._src.input_checks import check_excitations
 from magpylib._src.style import DefaultMarkers
@@ -87,8 +88,6 @@ class MagpyMarkers:
 
 def make_DefaultTrace(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     **kwargs,
@@ -111,15 +110,11 @@ def make_DefaultTrace(
     }
     update_trace_name(trace, f"{type(obj).__name__}", "", style)
     trace["text"] = trace["name"]
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Line(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     **kwargs,
@@ -153,15 +148,11 @@ def make_Line(
         else " (Current not initialized)"
     )
     update_trace_name(trace, "Line", default_suffix, style)
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Loop(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     vertices=50,
@@ -192,15 +183,11 @@ def make_Loop(
         else " (Current not initialized)"
     )
     update_trace_name(trace, "Loop", default_suffix, style)
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Dipole(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     autosize=None,
@@ -237,10 +224,88 @@ def make_Dipole(
     t = np.arccos(dot)
     vec = -t * cross / n
     mag_orient = RotScipy.from_rotvec(vec)
-    orientation = orientation * mag_orient
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    trace = place_and_orient_model3d(trace, orientation=mag_orient, **kwargs)
+    return {**trace, **kwargs}
+
+
+def make_mesh_lines(
+    obj,
+    pos_orient_inds,
+    mode,
+    label=None,
+    style=None,
+    color=None,  # pylint: disable=unused-argument
+    **kwargs,
+):
+    """Draw open or self-intersecting mesh lines and vertices"""
+    # pylint: disable=protected-access
+    style = obj.style if style is None else style
+    mesh = getattr(style.mesh, mode)
+    marker, line = mesh.marker, mesh.line
+    tr, vert = obj.faces, obj.vertices
+    if mode == "disconnected":
+        subsets = obj.get_faces_subsets()
+        lines = get_closest_vertices(subsets, vert)
+    else:
+        edges = np.concatenate([tr[:, 0:2], tr[:, 1:3], tr[:, ::2]], axis=0)
+        # make sure unique pairs are found regardless of order
+        edges = np.sort(edges, axis=1)
+        edges_uniq, edges_counts = np.unique(edges, axis=0, return_counts=True)
+        if mode == "open":
+            lines = vert[edges_uniq[edges_counts != 2]]
+        else:
+            lines = vert[edges_uniq]
+
+    out = {}
+    if lines.size != 0:
+        label = f"{obj}" if label is None else label
+        lines = np.insert(lines, 2, None, axis=1).reshape(-1, 3)
+        traces = []
+        for ind in pos_orient_inds:
+            x, y, z = (obj._orientation[ind].apply(lines) + obj._position[ind]).T
+            trace = {
+                "type": "scatter3d",
+                "x": x,
+                "y": y,
+                "z": z,
+                "marker_color": marker.color,
+                "marker_size": marker.size,
+                "marker_symbol": marker.symbol,
+                "line_color": line.color,
+                "line_width": line.width,
+                "line_dash": line.style,
+                "legendgroup": f"{obj}{mode}edges",
+                "name": f"{label} - {mode}-edges",
+            }
+            traces.append(trace)
+        out = {**merge_traces(*traces), **kwargs}
+    return out
+
+
+def get_closest_vertices(faces_subsets, vertices):
+    """Get closest pairs of points between disconnected subsets of faces indices"""
+    nparts = len(faces_subsets)
+    inds_subsets = [np.unique(v) for v in faces_subsets]
+    closest_verts_list = []
+    if nparts > 1:
+        connected = [np.min(inds_subsets[0])]
+        while len(connected) < nparts:
+            prev_min = float("inf")
+            for i in connected:
+                for j in range(nparts):
+                    if j not in connected:
+                        tr1, tr2 = inds_subsets[i], inds_subsets[j]
+                        c1, c2 = vertices[tr1], vertices[tr2]
+                        dist = distance.cdist(c1, c2)
+                        i1, i2 = divmod(dist.argmin(), dist.shape[1])
+                        min_dist = dist[i1, i2]
+                        if min_dist < prev_min:
+                            prev_min = min_dist
+                            closest_verts = [c1[i1], c2[i2]]
+                            connected_ind = j
+            connected.append(connected_ind)
+            closest_verts_list.append(closest_verts)
+    return np.array(closest_verts_list)
 
 
 def make_mesh_lines(
@@ -382,8 +447,6 @@ def make_triangle_orientations(
 
 def make_Cuboid(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     **kwargs,
@@ -398,18 +461,11 @@ def make_Cuboid(
     trace = make_BaseCuboid("plotly-dict", dimension=dimension, color=color)
     default_suffix = f" ({d[0]}m|{d[1]}m|{d[2]}m)"
     update_trace_name(trace, "Cuboid", default_suffix, style)
-    update_magnet_mesh(
-        trace, mag_style=style.magnetization, magnetization=obj.magnetization
-    )
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Cylinder(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     base=50,
@@ -427,18 +483,11 @@ def make_Cylinder(
     )
     default_suffix = f" (D={d[0]}m, H={d[1]}m)"
     update_trace_name(trace, "Cylinder", default_suffix, style)
-    update_magnet_mesh(
-        trace, mag_style=style.magnetization, magnetization=obj.magnetization
-    )
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_CylinderSegment(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     vertices=25,
@@ -456,18 +505,11 @@ def make_CylinderSegment(
     )
     default_suffix = f" (r={d[0]}m|{d[1]}m, h={d[2]}m, φ={d[3]}°|{d[4]}°)"
     update_trace_name(trace, "CylinderSegment", default_suffix, style)
-    update_magnet_mesh(
-        trace, mag_style=style.magnetization, magnetization=obj.magnetization
-    )
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Sphere(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     vertices=15,
@@ -485,18 +527,11 @@ def make_Sphere(
     )
     default_suffix = f" (D={unit_prefix(diameter / 1000)}m)"
     update_trace_name(trace, "Sphere", default_suffix, style)
-    update_magnet_mesh(
-        trace, mag_style=style.magnetization, magnetization=obj.magnetization
-    )
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Tetrahedron(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     **kwargs,
@@ -508,18 +543,11 @@ def make_Tetrahedron(
     style = obj.style if style is None else style
     trace = make_BaseTetrahedron("plotly-dict", vertices=obj.vertices, color=color)
     update_trace_name(trace, "Tetrahedron", "", style)
-    update_magnet_mesh(
-        trace, mag_style=style.magnetization, magnetization=obj.magnetization
-    )
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
 def make_Triangle(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     **kwargs,
@@ -551,12 +579,29 @@ def make_Triangle(
         "plotly-dict", vertices=vert, faces=faces, color=color
     )
     update_trace_name(trace, obj.__class__.__name__, "", style)
-    update_magnet_mesh(
-        trace, mag_style=style.magnetization, magnetization=obj.magnetization
+    return {**trace, **kwargs}
+
+
+def make_TriangularMesh(
+    obj,
+    color=None,
+    style=None,
+    **kwargs,
+) -> dict:
+    """
+    Creates the plotly mesh3d parameters for a Trianglular facet mesh in a dictionary based on the
+    provided arguments.
+    """
+    style = obj.style if style is None else style
+    trace = make_BaseTriangularMesh(
+        "plotly-dict", vertices=obj.vertices, faces=obj.faces, color=color
     )
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    ntri = len(obj.faces)
+    default_suffix = f" ({ntri} face{'s'[:ntri^1]})"
+    update_trace_name(trace, obj.__class__.__name__, default_suffix, style)
+    # make edges sharper in plotly
+    trace.update(flatshading=True, lighting_facenormalsepsilon=0, lighting_ambient=0.7)
+    return {**trace, **kwargs}
 
 
 def make_TriangularMesh(
@@ -602,8 +647,6 @@ def make_Pixels(positions, size=1) -> dict:
 
 def make_Sensor(
     obj,
-    position=(0.0, 0.0, 0.0),
-    orientation=None,
     color=None,
     style=None,
     autosize=None,
@@ -679,15 +722,17 @@ def make_Sensor(
         else ""
     )
     update_trace_name(trace, "Sensor", default_suffix, style)
-    return place_and_orient_model3d(
-        trace, orientation=orientation, position=position, **kwargs
-    )
+    return {**trace, **kwargs}
 
 
-def update_magnet_mesh(mesh_dict, mag_style=None, magnetization=None):
+def update_magnet_mesh(
+    mesh_dict, mag_style=None, magnetization=None, color_slicing=False
+):
     """
     Updates an existing plotly mesh3d dictionary of an object which has a magnetic vector. The
     object gets colorized, positioned and oriented based on provided arguments.
+    Slicing allows for matplotlib to show colorgradients approximations by slicing the mesh into
+    the colorscales colors, remesh it and merge with assigning facecolor for each part.
     """
     mag_color = mag_style.color
     if magnetization is not None and mag_style.show:
@@ -697,17 +742,24 @@ def update_magnet_mesh(mesh_dict, mag_style=None, magnetization=None):
             color_middle = mesh_dict["color"]
         elif mag_color.mode == "bicolor":
             color_middle = False
-        mesh_dict["colorscale"] = getColorscale(
-            color_transition=mag_color.transition,
+        ct = mag_color.transition
+        cs = getColorscale(
+            color_transition=0 if color_slicing else ct,
             color_north=mag_color.north,
             color_middle=color_middle,
             color_south=mag_color.south,
         )
-        mesh_dict["intensity"] = getIntensity(
-            vertices=vertices,
-            axis=magnetization,
-        )
+        if color_slicing:
+            tr = slice_mesh_from_colorscale(mesh_dict, magnetization, cs)
+            mesh_dict.update(tr)
+        else:
+            mesh_dict["colorscale"] = cs
+            mesh_dict["intensity"] = getIntensity(
+                vertices=vertices,
+                axis=magnetization,
+            )
         mesh_dict["showscale"] = False
+        mesh_dict.pop("color_slicing", None)
     return mesh_dict
 
 
@@ -850,13 +902,13 @@ def get_generic_traces(
     legendgroup = f"{input_obj}" if legendgroup is None else legendgroup
 
     is_mag_arrows = False
-    if getattr(input_obj, "magnetization", None) is not None:
-        mode = style.magnetization.mode
-        if style.magnetization.show:
-            if "arrow" in mode or not mag_color_grad_apt:
-                is_mag_arrows = True
-            if mag_color_grad_apt and "color" not in mode and mode != "auto":
-                style.magnetization.show = False  # disables color gradient only
+    is_mag = hasattr(input_obj, "magnetization")
+    if is_mag and style.magnetization.show:
+        mag = style.magnetization
+        if mag.mode == "auto":
+            mag.mode = "color"  # if mag_color_grad_apt else "arrow"
+        is_mag_arrows = "arrow" in mag.mode
+        mag.show = "color" in mag.mode
 
     # check excitations validity
     for param in ("magnetization", "arrow"):
@@ -887,11 +939,34 @@ def get_generic_traces(
     orientations, positions, pos_orient_inds = get_rot_pos_from_path(
         input_obj, style.path.frames
     )
-    obj_is_disconnected = (
-        isinstance(input_obj, TriangularMesh)
-        and style.mesh.disconnected.show
-        and input_obj.check_disconnected()
-    )
+    obj_is_disconnected = False
+    if isinstance(input_obj, TriangularMesh):
+        for mode in ("open", "disconnected"):
+            show_mesh = getattr(style.mesh, mode).show
+            if mode == "open" and show_mesh:
+                if input_obj.status_open is None:
+                    warnings.warn(
+                        f"Unchecked open mesh status in {input_obj!r} detected, before attempting "
+                        "to show potential open edges, which may take a while to compute "
+                        "when the mesh has many faces, now applying operation..."
+                    )
+                    input_obj.check_open()
+            elif mode == "disconnected" and show_mesh:
+                if input_obj.status_disconnected is None:
+                    warnings.warn(
+                        f"Unchecked disconnected mesh status in {input_obj!r} detected, before "
+                        "attempting to show possible disconnected parts, which may take a while "
+                        "to compute when the mesh has many faces, now applying operation..."
+                    )
+                obj_is_disconnected = input_obj.check_disconnected()
+            if mode == "selfintersecting":
+                if input_obj._status_selfintersecting is None:
+                    warnings.warn(
+                        f"Unchecked{input_obj!r} selfintersecting status has not been checked before "
+                        "atempting to show possible self-intersecting parts, which may take a "
+                        "while to compute when the mesh has many faces, now applying operation..."
+                    )
+                    input_obj.check_selfintersecting()
     disconnected_traces = []
     for pos_orient_enum, (orient, pos) in enumerate(zip(orientations, positions)):
         if style.model3d.showdefault and make_func is not None:
@@ -905,19 +980,26 @@ def get_generic_traces(
                     # temporary mutate faces from subset
                     input_obj._faces = tri
                     style.magnetization.show = False
-                    disconnected_traces.append(
-                        make_func(
-                            position=pos,
-                            orientation=orient,
-                            **{**make_func_kwargs, "color": dis_color},
-                        )
+                    dis_tr = make_func(
+                        **{**make_func_kwargs, "color": dis_color},
                     )
+                    dis_tr = place_and_orient_model3d(
+                        dis_tr, orientation=orient, position=pos
+                    )
+                    disconnected_traces.append(dis_tr)
                 input_obj._faces = tria_orig
                 style.magnetization.show = mag_show
-            else:
-                path_traces.append(
-                    make_func(position=pos, orientation=orient, **make_func_kwargs)
-                )
+            else:  # if disconnnected, no mag slicing needed
+                p_tr = make_func(**make_func_kwargs)
+                if is_mag:
+                    p_tr = update_magnet_mesh(
+                        p_tr,
+                        mag_style=style.magnetization,
+                        magnetization=input_obj.magnetization,
+                        color_slicing=not mag_color_grad_apt,
+                    )
+                p_tr = place_and_orient_model3d(p_tr, orientation=orient, position=pos)
+                path_traces.append(p_tr)
         for extr in extra_model3d_traces:
             if extr.show:
                 extr.update(extr.updatefunc())
@@ -1029,33 +1111,6 @@ def get_generic_traces(
         )
     if isinstance(input_obj, TriangularMesh):
         for mode in ("grid", "open", "disconnected", "selfintersecting"):
-            if mode == "open":
-                if input_obj._status_open is None:
-                    warnings.warn(
-                        f"{input_obj!r} open status has not been checked before attempting "
-                        "to show potential open edges, which may take a while to compute "
-                        "when the mesh has many faces, now applying operation..."
-                    )
-                if not input_obj.check_open():
-                    continue
-            if mode == "disconnected":
-                if input_obj._status_disconnected is None:
-                    warnings.warn(
-                        f"{input_obj!r} disconnected status has not been checked before atempting "
-                        "to show possible disconnected parts, which may take a while to compute "
-                        "when the mesh has many faces, now applying operation..."
-                    )
-                if not input_obj.check_disconnected():
-                    continue
-            if mode == "selfintersecting":
-                if input_obj._status_selfintersecting is None:
-                    warnings.warn(
-                        f"{input_obj!r} selfintersecting status has not been checked before "
-                        "atempting to show possible self-intersecting parts, which may take a "
-                        "while to compute when the mesh has many faces, now applying operation..."
-                    )
-                if not input_obj.check_selfintersecting():
-                    continue
             if getattr(style.mesh, mode).show:
                 trace = make_mesh_lines(
                     input_obj, pos_orient_inds, mode, label, **kwargs